Move all back to TracyProfiler.cpp. Only need the fix in HandleSymbolQuery, not DecodeSymbolAddress
